LPD-15801 Refactor Fields to Table
