+++ modules/apps/frontend-data-set/frontend-data-set-views-web/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/list/List.tsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 
-import {IBaseVisualizationMode} from './VisualizationModes';
+import {IBaseVisualizationMode} from '../VisualizationModes';
 
 export interface IList extends IBaseVisualizationMode<'list'> {}
